requirements.txt
python-dotenv==0.15.0
requests~=2.25.1
python-telegram-bot==13.3
redis==3.5.3